+++ NEWS.md
@@ -1,7 +1,8 @@
 
-# rquery 1.2.1 2018/11/07
+# rquery 1.2.1 2018/11/11
 
  * versions without _nse() suffix.
+ * parse ^ power.
  * update docs.
 
 # rquery 1.2.0 2018/11/06